updated get recommendations filter api and other related model field modificaitron
