+++ adapt/common/models/house.js
@@ -1,7 +1,16 @@
 'use strict';
 var uniq = require('lodash.uniq');//unique eleemnts in array
+var union = require('lodash.union');
 
 module.exports = function(House) {
+
+  House.remoteMethod('getRecommendations',{
+    accepts: [{arg: 'id', type: 'number', required: true},
+    {arg: 'roomType', type: 'number'}],
+    http: {path: '/:id/getRecommendations/:roomType', verb: 'get'},
+    returns: { type: ['object'], root: true }
+  });
+
   House.observe('before save',async function(ctx) {
     var houseInstance = ctx.instance;
     //get the residents list for this house
@@ -11,38 +20,82 @@ module.exports = function(House) {
     return;
   });
 
-  //go through all residents of the house and
-  House.getFilterableCategories = function(houseID, cb) {
-    return House.findById(houseID,{include: {relation: 'residents'}}, function(err, house){
 
-      var categories = {
-        "incomes":[],
-        "ages":[],
-        "genders":[],
-        "locations":[],
-        "concerns":[]
-      }
-      if (house == null){
-        return categories;
-      }
-      var incomes = [];
-      var ages = [];
-      var genders = [];
-      var locations = [];
-      var concerns = [];
-      house.toJSON().residents.forEach(function(resident){
-        categories.incomes.push(resident.income);
-        categories.ages.push(resident.age);
-        categories.genders.push(resident.gender);
-        categories.locations.push(resident.location);
-      });
-      categories.incomes = uniq(categories.incomes);
-      categories.ages = uniq(categories.ages);
-      categories.genders = uniq(categories.genders);
-      categories.locations = uniq(categories.locations);
+  House.getRecommendations = function(houseId, roomType,cb) {
+    //https://loopback.io/doc/en/lb3/Querying-data.html
+    var app = House.app;
 
-      cb(null, categories);
+    House.findById(houseId,{include: {relation: 'residents'}}, function(err, house){
+      if (err == null && house != null) {
+      
+        const categories = House.getFilterableCategories(house);
+        var Recommendation = app.models.Recommendation;
+        console.log(categories);
+  
+        Recommendation.find(function(err,recs){
+          var filtered = recs;
+          filtered = filtered.filter(rec => {
+            return rec.applicableRoomTypes.includes(roomType);
+          })
+          filtered = filtered.filter(rec => {
+            const intersection = categories.incomes.filter(income=>rec.applicableIncomes.includes(income));
+            return intersection.length > 0;
+          })
+
+          filtered = filtered.filter(rec => {
+            const intersection = categories.ages.filter(age=>rec.applicableAges.includes(age));
+            return intersection.length > 0;
+          })
+
+          filtered = filtered.filter(rec => {
+            const intersection = categories.genders.filter(gender=>rec.applicableGenders.includes(gender));
+            return intersection.length > 0;
+          })
+          filtered = filtered.filter(rec => {
+            const intersection = categories.locations.filter(state=>rec.applicableLocations.includes(state));
+            return intersection.length > 0;
+          })
+          filtered = filtered.filter(rec => {
+            const intersection = categories.concerns.filter(concern=>rec.applicableConcerns.includes(concern));
+            return intersection.length > 0;
+          })
+          cb(null,filtered);
+        });
+      }else{
+        console.log(err);
+        cb(err,[]);
+      }
     });
+
+  }
+
+
+  //go through all residents of the house and
+  House.getFilterableCategories = function(house) {
+    var categories = {
+      "incomes":[],
+      "ages":[],
+      "genders":[],
+      "locations":[],
+      "concerns":[]
+    }
+    if (house == null){
+      return categories;
+    }
+    house.toJSON().residents.forEach(function(resident){
+      categories.incomes.push(resident.income);
+      categories.ages.push(resident.age);
+      categories.genders.push(resident.gender);
+      categories.locations.push(resident.location);
+      categories.concerns =  union(resident.concerns, categories.concerns);
+    });
+    categories.incomes = uniq(categories.incomes);
+    categories.ages = uniq(categories.ages);
+    categories.genders = uniq(categories.genders);
+    categories.locations = uniq(categories.locations);
+    //categories.concerns = uniq(categories.concerns);
+
+    return categories;
 
   }
