+++ adapt/package.json
@@ -15,6 +15,7 @@
     "compression": "^1.0.3",
     "cors": "^2.5.2",
     "helmet": "^3.10.0",
+    "lodash.union": "^4.6.0",
     "lodash.uniq": "^4.5.0",
     "loopback": "^3.22.0",
     "loopback-boot": "^2.6.5",
