+++ docker-entrypoint.sh
@@ -7,6 +7,6 @@ echo "changing into app directory"
 cd adapt
 sudo npm install
 echo "migrating models from loopback to database"
-node ./server/create-lb-tables.js
+node ./create-lb-tables.js
 echo "running adapt on local server"
 node .
